refactor: Update component imports in LoginLogo and VerticalMenuHeader
