+++ src/components/atoms/PortalLogo.tsx
